파일 위치 변경 및 네임스페이스 삭제
