+++ Example/main.cpp
@@ -1,0 +1,6 @@
+#include "Enviroment.h"
+
+int main() {
+    
+    
+}
