+++ include/Graphics/Enviroment.h
@@ -1,6 +1,5 @@
 #pragma once
 
-
 // OS
 #if defined(__APPLE__) && !defined(Enviroment_OS)
 	#define Enviroment_OS OperatingSystem::MacOS;
@@ -27,7 +26,6 @@
 	#define OS_LINUX false
 	#define OS_OTHERS true
 #endif
-
 
 // Compiler
 #if defined(_MSC_VER) && !defined(Enviroment_Compiler)
@@ -80,3 +78,33 @@
 	#define ARCHITECTURE_X64 false
 	#define ARCHITECTURE_X86 true
 #endif
+
+#if OS_WINDOWS
+#include <Windows.h>
+#endif
+
+namespace Graphics{
+	namespace Enviroment {
+		enum class OperatingSystem {
+			Windows,
+			Linux,
+			MacOS,
+			Others
+		};
+		enum class KindCompiler {
+			MSVC,
+			MingW,
+			GNUC,
+			Others
+		};
+
+		enum class KindArchitecture {
+			X86,
+			X64
+		};
+
+		constexpr OperatingSystem OS = Enviroment_OS;
+		constexpr KindCompiler Compiler = Enviroment_Compiler;
+		constexpr KindArchitecture Architecture = Enviroment_Architecture;
+	}
+}
