+++ lib/windows/Graphics/Mouse.cpp
@@ -3,7 +3,7 @@
 #include "Mouse.h"
 
 
-void Graphics::Input::Mouse::KeyDown(MouseButton id, Point pos)
+void Graphics::Mouse::KeyDown(MouseButton id, Point pos)
 {
     if (event[(int)id].Position == pos && event[(int)id].GetMouseButton() == id)
     {
@@ -17,7 +17,7 @@ void Graphics::Input::Mouse::KeyDown(MouseButton id, Point pos)
     EventKeyChanged.Invoke(event, BUTTON_SIZE);
 }
 
-void Graphics::Input::Mouse::KeyUp(MouseButton id, Point pos)
+void Graphics::Mouse::KeyUp(MouseButton id, Point pos)
 {
     event[(int)id].Position = pos;
     event[(int)id].Pressed = false;
@@ -26,7 +26,7 @@ void Graphics::Input::Mouse::KeyUp(MouseButton id, Point pos)
     EventKeyChanged.Invoke(event, BUTTON_SIZE);
 }
 
-void Graphics::Input::Mouse::Refresh(void *data)
+void Graphics::Mouse::Refresh(void *data)
 {
     MOUSE_EVENT_RECORD input = *(MOUSE_EVENT_RECORD *)data;
