--- prev/UnitTest/GraphicsTest.h
@@ -1,12 +1,0 @@
-#pragma once
-
-
-#include "EventHandlerTest.h"
-#include "FunctionTest.h"
-
-void GrahpicsTest() {
-	EventHanlderTest();
-    FunctionTest();
-    DisplayResizeTest();
-
-}
