--- prev/include/Graphics/IDevice.h
@@ -1,13 +1,0 @@
-#pragma once
-
-
-namespace Graphics {
-    namespace Library {
-        namespace Interface {
-            class IDevice {
-				virtual void Refresh(void*) = 0;
-
-            };
-        }
-    }
-}
