--- prev/include/Graphics/Shape/Circle.h
@@ -1,1 +1,0 @@
-#pragma once
